requirements.txt
dash==2.9.3
pandas==2.0.1
plotly-express==0.4.1